fix: repair the debug build after RSet::valid became atomic
519c6e23a made RSet::valid a std::atomic_bool but left two readers that only
compile when DEBUG is defined, so the break is invisible to a release build and
only shows up where assertions are live:

- RSet::isValid() used "rs ? rs->valid : false".  The two arms are now unrelated
  types, each convertible to the other, which makes the conditional ambiguous;
  use "rs && rs->valid" instead, which reads the atomic through the same
  implicit conversion every other reader uses.

- RSet::canDelete()'s entry printd() passed the atomic itself to a variadic,
  whose copy constructor is deleted.  Cast to int, as the sibling printd() in
  RSet::dbg() already does.

Neither changes behavior; both are debug-only lines.

Tests: cmake --build build-debug (was failing to compile, now clean);
run_unit_tests(): 671 tests, 0 failures; gc, recursive, dgc-scan-restart,
dgc-closed-sets, dgc-unchanged-sets, dgc-graph-components and
dgc-opaque-references all pass in build-debug

+++ lib/RSet.cpp
@@ -484,7 +484,7 @@ void RSet::dbg() {
 
 // if we return 1, the rset has been invalidated already
 int RSet::canDelete(int ref_copy, int rcount, int scan_refs, RObject& initiator, RSetDerefHelper& cleanup) {
-    printd(QRO_LVL, "RSet::canDelete() this: %p valid: %d\n", this, valid);
+    printd(QRO_LVL, "RSet::canDelete() this: %p valid: %d\n", this, (int)valid);
 
     if (q_disable_gc)
         return 0;
